applications/bot/pyproject.toml
[tool.poetry]
name = "Saltyboy Bot"
version = "1.0.0"
description = "Bot to scrape Twitch's IRC for SaltyBet and store fighters and matches into a PostgreSQL database."
authors = ["how2vim <francisco.ad.trindade@gmail.com> & diegowsan"]
license = "MIT"
package-mode = false 

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
scikit-learn = "^1.7.2"
pandas = "^2.3.3"
numpy = "^2.3.5"
requests = "^2.32.5"
tenacity = "^9.1.2"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
mypy = "^1.8.0"
pycln = "^2.4.0"
pylint = "^3.0.3"
pylint-exit = "^1.2.0"
ruff = "^0.2.2"
types-psycopg2 = "^2.9.21.20240218"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pylint.format]
max-line-length = "88"

[tool.pylint.MASTER]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-many-arguments",
    "too-many-instance-attributes",
    "broad-exception-caught",
    "invalid-name",
    "line-too-long",
]
ignored-modules = ["alembic", "sqlalchemy"]
